# docs/swagger.yaml
basePath: /api
definitions:
...
  httpresp.HTTPError:
    properties:
      code:
        description: the HTTP status
        example: 500
        type: integer
      error:
        description: the error description
        example: internal server error
        type: string
      success:
        description: wether the business action is successfull or not
        example: false
        type: boolean
    type: object
  httpresp.HTTPError401:
    properties:
      code:
        description: the HTTP status
        example: 401
        type: integer
      error:
        description: the error description
        example: unauthorized
        type: string
      success:
        description: wether the business action is successfull or not
        example: false
        type: boolean
    type: object
  httpresp.JSONResult:
    properties:
      code:
        description: the HTTP status
        example: 200
        type: integer
      result:
        description: whatever object or array of objects
      success:
        description: wether the business action is successfull or not
        example: true
        type: boolean
    type: object
...
  interfaces.Credential:
    properties:
      password:
        example: "1234"
        type: string
      username:
        example: "042"
        type: string
    type: object
...
  interfaces.Token:
    properties:
      accessToken:
        description: token
        type: string
      expiration:
        description: expiration date
        example: ""
        type: string
      refreshToken:
        description: refresh token to be used for renewal
        type: string
    type: object
...
host: localhost:5001
info:
  contact:
    email: malek.radhouen@gmail.com
    name: API Support
    url: https://www.magic.fr/support
  description: Here is our solution documentation and testing portal of provided functionalities
    to interact with our hypervisor tool.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: https://www.magic.com/terms/
  title: Magic - Inventory solution API
  version: "1.0"
paths:
  /activate/{token}:
    get:
      consumes:
      - application/json
      description: Activates a user account using the validation token from email
      operationId: activate-account
      parameters:
      - description: Activation token
        in: path
        name: token
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Account activated successfully
          schema:
            $ref: '#/definitions/httpresp.JSONResult'
        "400":
          description: Invalid or expired token
          schema:
            $ref: '#/definitions/httpresp.HTTPError'
      summary: Activate user account
      tags:
      - authenticate
...
  /authenticate:
    post:
      consumes:
      - application/json
      description: authenticate user using provided credentials. Returned token can
        be used as "Bearer [insert token here]" in the 'Authorize' form
      operationId: authenticate-user-by-username-password
      parameters:
      - description: Authentication credentials
        in: body
        name: credential
        required: true
        schema:
          $ref: '#/definitions/interfaces.Credential'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/httpresp.JSONResult'
            - properties:
                result:
                  $ref: '#/definitions/interfaces.Token'
                success:
                  type: boolean
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httpresp.HTTPError401'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httpresp.HTTPError'
      summary: Authenticate user
      tags:
      - authenticate
...
  /forgot-password:
    post:
      consumes:
      - application/json
      description: Sends a password reset email to the user
      operationId: forgot-password
      parameters:
      - description: Email address
        in: body
        name: email
        required: true
        schema:
          additionalProperties:
            type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: Password reset email sent
          schema:
            $ref: '#/definitions/httpresp.JSONResult'
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/httpresp.HTTPError'
      summary: Request password reset
      tags:
      - authenticate
  /logout:
    post:
      consumes:
      - application/json
      description: Logs out the authenticated user by invalidating the token
      operationId: logout-user
      produces:
      - application/json
      responses:
        "200":
          description: Successfully logged out
          schema:
            $ref: '#/definitions/httpresp.JSONResult'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httpresp.HTTPError401'
      security:
      - ApiKeyAuth: []
      summary: Logout user
      tags:
      - authenticate
...
  /refresh_token:
    post:
      consumes:
      - application/json
      description: refreshes the provided token
      operationId: refresh-token
      parameters:
      - description: Token
        in: query
        name: token
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/httpresp.JSONResult'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httpresp.HTTPError401'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httpresp.HTTPError'
      security:
      - ApiKeyAuth: []
      summary: RefreshToken refreshes authenticated user's token
      tags:
      - authenticate
  /reset-password:
    post:
      consumes:
      - application/json
      description: Resets the user's password using the token
      operationId: reset-password
      parameters:
      - description: Token and new password
        in: body
        name: request
        required: true
        schema:
          additionalProperties:
            type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: Password reset successfully
          schema:
            $ref: '#/definitions/httpresp.JSONResult'
        "400":
          description: Invalid or expired token
          schema:
            $ref: '#/definitions/httpresp.HTTPError'
      summary: Reset password
      tags:
      - authenticate
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"
...